+++ src/utils/sanitize.py
@@ -1,2 +1,7 @@
 def sanitize(data: str) -> str:
-    return data
+    result = ''
+    for char in data:
+        if 'a' <= char <= 'z' or 'A' <= char <= 'Z' or '0' <= char <= '9':
+            result += char
+
+    return result