+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ## 0.0.8 (2020-xx-xx Unreleased)
 
-- coming soon...
+- add test-cases for sanitize
 
 ## 0.0.7 (2020-01-26)